+++ apps/www/src/components/Chat.tsx
@@ -2,6 +2,7 @@ import { Message as MessageComponent } from "@/components/demo/message/Message";
 import { Button } from "@/components/ui/Button";
 import { Input } from "@/components/ui/Input";
 import { Textarea } from "@/components/ui/Textarea";
+import { Icons } from "@/icons";
 import { nanoid, type Message } from "ai";
 import { useChat } from "ai/react";
 import * as React from "react";
@@ -10,13 +11,22 @@ const ChatMessage: React.FC<{
   message: Message;
   isStreamFinished: boolean;
 }> = ({ message, isStreamFinished }) => {
+  const { role, content } = message;
   return (
-    <div>
-      <p>{message.role}</p>
-      <MessageComponent
-        message={message.content}
-        isStreamFinished={isStreamFinished}
-      />
+    <div className="group relative flex items-start m-2">
+      <div className="bg-background flex size-[25px] shrink-0 select-none items-center justify-center rounded-md border shadow-sm">
+        {role === "user" ? (
+          <Icons.user className="size-4" />
+        ) : (
+          <Icons.bot className="size-4" />
+        )}
+      </div>
+      <div className="ml-4 flex-1 space-y-2 overflow-hidden">
+        <MessageComponent
+          message={content}
+          isStreamFinished={isStreamFinished}
+        />
+      </div>
     </div>
   );
 };
@@ -45,32 +55,59 @@ export const Chat = () => {
 
   const messagesWithoutSystem = messages.slice(1);
   return (
-    <div>
-      <form autoComplete="off" onSubmit={handleSubmit} className="space-y-8">
-        <label>Enter Your API Key</label>
-        <Input value={currentApiKey} onChange={handleUpdateApiKey} />
-        {messagesWithoutSystem.map((message, index) => {
-          const isStreamFinished =
-            ["user", "system"].includes(message.role) ||
-            index < messagesWithoutSystem.length - 1 ||
-            !isLoading;
-          return (
-            <ChatMessage
-              key={message.id}
-              message={message}
-              isStreamFinished={isStreamFinished}
-            />
-          );
-        })}
-        <Textarea
-          placeholder="What would you like to know?"
-          value={input}
-          onChange={handleInputChange}
-          className="mb-2"
+    <div className="bg-muted/50 relative w-full min-h-[calc(100vh-theme(spacing.18))]">
+      <div className="absolute top-4 left-4">
+        <Input
+          value={currentApiKey}
+          className="focus-within:border-white"
+          placeholder="Enter Your API Key"
+          onChange={handleUpdateApiKey}
         />
-        <Button disabled={isLoading} type="submit">
-          Start
-        </Button>
+      </div>
+      <form
+        autoComplete="off"
+        onSubmit={handleSubmit}
+        className="p-2 flex flex-col max-w-2xl mx-auto"
+      >
+        <div className="pb-[200px] pt-4 md:pt-20">
+          {messagesWithoutSystem.map((message, index) => {
+            const isStreamFinished =
+              ["user", "system"].includes(message.role) ||
+              index < messagesWithoutSystem.length - 1 ||
+              !isLoading;
+            return (
+              <div key={message.id}>
+                <ChatMessage
+                  key={message.id}
+                  message={message}
+                  isStreamFinished={isStreamFinished}
+                />
+                {index != messagesWithoutSystem.length - 1 && (
+                  <div className="shrink-0 bg-border h-[1px] w-full my-4"></div>
+                )}
+              </div>
+            );
+          })}
+        </div>
+        <div className="w-full fixed bottom-0">
+          <div className="sm:max-w-2xl">
+            <div className="bg-background flex flex-col overflow-hidden max-h-60 focus-within:border-white relative px-4 py-2 shadow-lg mb-2 sm:rounded-xl sm:border md:py-4">
+              <Textarea
+                placeholder="What would you like to know?"
+                value={input}
+                onChange={handleInputChange}
+                className="min-h-[60px] w-[calc(100%-theme(spacing.18))] focus-visible:ring-0 resize-none bg-transparent focus-within:outline-none sm:text-base border-none"
+              />
+              <Button
+                disabled={isLoading || !input}
+                className="absolute right-0 bottom-2 sm:right-4"
+                type="submit"
+              >
+                Run <Icons.return className="size-4 ml-2" />
+              </Button>
+            </div>
+          </div>
+        </div>
       </form>
     </div>
   );