+++ apps/www/src/icons/index.tsx
@@ -1,8 +1,11 @@
 import {
+  Bot,
   Calendar,
   Clock,
+  CornerDownLeft,
   Flame,
   Tv2,
+  User,
   type LucideIcon,
   type LucideProps,
 } from "lucide-react";
@@ -14,6 +17,9 @@ export const Icons = {
   calendar: Calendar,
   tv: Tv2,
   clock: Clock,
+  user: User,
+  bot: Bot,
+  return: CornerDownLeft,
   logo: ({ ...props }: LucideProps) => (
     <svg
       xmlns="http://www.w3.org/2000/svg"