+++ examples/custom-components/nextjs/.next/server/server-reference-manifest.js
@@ -1,0 +1,1 @@
+self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"OMWhi1NDyxZRpRPRD5H2JyZ/OfI93TW5T7s4gUEBX60=\"\n}"