+++ packages/custom/src/shared.ts
@@ -1,8 +1,0 @@
-import { CustomBlockOptions } from "./options";
-
-export const removeStartEndChars = (
-  str: string,
-  { startChar, endChar }: CustomBlockOptions,
-) => {
-  return str.replaceAll(startChar, "").replaceAll(endChar, "");
-};