+++ packages/markdown/src/markdownParser.ts
@@ -200,7 +200,10 @@ const markdownAstToVisibleText = (markdownAst: Root, isFinished: boolean) => {
   if (!isFinished) {
     removePartialAmbiguousMarkdownFromAst(markdownAst);
   }
-  return removeZeroWidthSpaces(markdownAstToVisibleTextHelper(markdownAst));
+  return removeZeroWidthSpaces(
+    markdownAstToVisibleTextHelper(markdownAst),
+  ).replaceAll("\n", "");
+  // mdast is not reliable with \n so we remove them all
 };
 
 export const markdownToVisibleText = (
@@ -290,7 +293,6 @@ export const markdownWithVisibleChars = (
     removePartialAmbiguousMarkdownFromAst(markdownAst);
   }
   const visibleText = markdownAstToVisibleText(markdownAst, isFinished);
-
   const charsToRemove = visibleText.length - visibleChars;
   removeVisibleCharsFromAst(markdownAst, charsToRemove);
   return astToMarkdown(markdownAst);